test(e2e): move remaining Detox confirmations to skipped Appium
Port gas-fee, per-dapp network, and transaction-pay specs as skipped
Appium suites, delete Detox twins, and retire Detox confirmations CI jobs.
