+++ .github/workflows/run-e2e-smoke-tests-android.yml
@@ -27,6 +27,9 @@ jobs:
   # SmokeWalletPlatform Detox job removed — tag fully migrated to Appium
   # (run-appium-smoke-tests-android.yml / appium-wallet-platform-android-smoke).
 
+  # SmokeConfirmations Detox job removed — tag fully migrated to Appium
+  # (run-appium-smoke-tests-android.yml / appium-confirmations-android-smoke).
+
   network-abstraction-android-smoke:
     if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokeNetworkAbstractions')
     strategy:
@@ -61,23 +64,6 @@ jobs:
       runner_provider: ${{ inputs.runner_provider }}
     secrets: inherit
 
-  confirmations-android-smoke:
-    if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokeConfirmations')
-    strategy:
-      matrix:
-        split: [1, 2, 3, 4]
-      fail-fast: false
-    uses: ./.github/workflows/run-e2e-workflow.yml
-    with:
-      test-suite-name: confirmations-android-smoke-${{ matrix.split }}
-      platform: android
-      test_suite_tag: 'SmokeConfirmations'
-      split_number: ${{ matrix.split }}
-      total_splits: 4
-      changed_spec_files: ${{ inputs.changed_spec_files }}
-      runner_provider: ${{ inputs.runner_provider }}
-    secrets: inherit
-
   report-android-smoke-tests:
     name: Report Android Smoke Tests
     runs-on: ${{ inputs.runner_provider == 'namespace' && 'namespace-profile-metamask-ci-linux' || 'ubuntu-latest' }}
@@ -94,7 +80,6 @@ jobs:
     needs:
       - network-abstraction-android-smoke
       - network-expansion-android-smoke
-      - confirmations-android-smoke
     steps:
       - name: Checkout (Namespace)
         uses: namespacelabs/nscloud-checkout-action@938f5d2d403d6224d9a0c0dc559b1dae09c2ede4 # v8.1.1
