+++ .github/workflows/run-e2e-smoke-tests-ios.yml
@@ -24,23 +24,8 @@ permissions:
   id-token: write
 
 jobs:
-  confirmations-ios-smoke:
-    if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokeConfirmations')
-    strategy:
-      matrix:
-        split: [1, 2, 3, 4]
-      fail-fast: false
-    uses: ./.github/workflows/run-e2e-workflow.yml
-    with:
-      test-suite-name: confirmations-ios-smoke-${{ matrix.split }}
-      platform: ios
-      test_suite_tag: 'SmokeConfirmations'
-      split_number: ${{ matrix.split }}
-      total_splits: 4
-      changed_spec_files: ${{ inputs.changed_spec_files }}
-      build_type: 'main'
-      runner_provider: ${{ inputs.runner_provider }}
-    secrets: inherit
+  # SmokeConfirmations Detox job removed — tag fully migrated to Appium
+  # (run-appium-smoke-tests-ios.yml / appium-confirmations-ios-smoke).
 
   # SmokeWalletPlatform Detox job removed — tag fully migrated to Appium
   # (run-appium-smoke-tests-ios.yml / appium-wallet-platform-ios-smoke).
@@ -95,7 +80,6 @@ jobs:
         )
       }}
     needs:
-      - confirmations-ios-smoke
       - network-abstraction-ios-smoke
       - network-expansion-ios-smoke
     steps:
