+++ tests/smoke-appium/confirmations/transactions/gas-fee-tokens-eip-7702.spec.ts
@@ -1,0 +1,352 @@
+import { test as appiumTest } from '../../../framework/fixtures/playwright/index.js';
+import FixtureBuilder, {
+  DEFAULT_FIXTURE_ACCOUNT,
+} from '../../../framework/fixtures/FixtureBuilder.js';
+import FooterActions from '../../../page-objects/Browser/Confirmations/FooterActions.js';
+import TabBarComponent from '../../../page-objects/wallet/TabBarComponent.js';
+import Assertions from '../../../framework/Assertions.js';
+import Matchers from '../../../framework/Matchers.js';
+import { LocalNode, LocalNodeType } from '../../../framework/types.js';
+import { SmokeConfirmations } from '../../../tags.js';
+import { loginToAppPlaywright } from '../../../flows/wallet.flow.js';
+import { startRedesignedNativeSendFiveEthToReview } from '../../../flows/confirmations.flow.js';
+import { withFixtures } from '../../../framework/fixtures/FixtureHelper.js';
+import RowComponents from '../../../page-objects/Browser/Confirmations/RowComponents.js';
+import { AnvilManager, Hardfork } from '../../../seeder/anvil-manager.js';
+import {
+  setupMockPostRequest,
+  setupMockRequest,
+} from '../../../api-mocking/helpers/mockHelpers.js';
+import { SIMULATION_ENABLED_NETWORKS_MOCK } from '../../../api-mocking/mock-responses/simulations.js';
+import { setupRemoteFeatureFlagsMock } from '../../../api-mocking/helpers/remoteFeatureFlagsHelper.js';
+import { remoteFeatureEip7702 } from '../../../api-mocking/mock-responses/feature-flags-mocks.js';
+import { Mockttp } from 'mockttp';
+import {
+  TRANSACTION_RELAY_STATUS_NETWORKS_MOCK,
+  TRANSACTION_RELAY_SUBMIT_NETWORKS_MOCK,
+} from '../../../api-mocking/mock-responses/transaction-relay-mocks.js';
+import { RelayStatus } from '../../../../app/util/transactions/transaction-relay.js';
+import TransactionConfirmView from '../../../page-objects/Send/TransactionConfirmView.js';
+import GasFeeTokenModal from '../../../page-objects/Confirmation/GasFeeTokenModal.js';
+import { AnvilPort } from '../../../framework/fixtures/FixtureUtils.js';
+
+const TRANSACTION_UUID_MOCK = '1234-5678';
+const SENDER_ADDRESS_MOCK = '0x76cf1cdd1fcc252442b50d6e97207228aa4aefc3';
+const RECIPIENT_ADDRESS_MOCK = '0x0c54fccd2e384b4bb6f2e405bf5cbc15a017aafb';
+const LOCALHOST_SENTINEL_URL =
+  'https://tx-sentinel-localhost.api.cx.metamask.io';
+
+const SEND_ETH_TRANSACTION_MOCK = {
+  data: '0x',
+  from: SENDER_ADDRESS_MOCK,
+  to: RECIPIENT_ADDRESS_MOCK,
+  value: '0xde0b6B3a7640000',
+};
+
+const SIMULATION_ENABLED_NETWORKS_WITH_RELAY = {
+  ...SIMULATION_ENABLED_NETWORKS_MOCK,
+  response: {
+    ...SIMULATION_ENABLED_NETWORKS_MOCK.response,
+    1337: {
+      ...SIMULATION_ENABLED_NETWORKS_MOCK.response[1337],
+      relayTransactions: true,
+      sendBundle: true,
+    },
+    1: {
+      network: 'ethereum-mainnet',
+      confirmations: true,
+      relayTransactions: true,
+      sendBundle: true,
+    },
+  },
+};
+
+const SIMULATION_GAS_STATION_MOCK = {
+  requestBody: {
+    jsonrpc: '2.0',
+    method: 'infura_simulateTransactions',
+    params: [
+      {
+        transactions: [SEND_ETH_TRANSACTION_MOCK],
+        suggestFees: {
+          withFeeTransfer: true,
+          withTransfer: true,
+          with7702: true,
+        },
+      },
+    ],
+  },
+  ignoreFields: [
+    'params.0.blockOverrides',
+    'id',
+    'params.0.transactions',
+    'params.0.suggestFees',
+  ],
+  urlEndpoint: LOCALHOST_SENTINEL_URL,
+  responseCode: 200,
+  response: {
+    jsonrpc: '2.0',
+    result: {
+      transactions: [
+        {
+          return:
+            '0x0000000000000000000000000000000000000000000000000000000000000000',
+          status: '0x1',
+          gasUsed: '0x5de2',
+          gasLimit: '0x5f34',
+          fees: [
+            {
+              maxFeePerGas: '0xf19b9f48d',
+              maxPriorityFeePerGas: '0x9febc9',
+              balanceNeeded: '0x59d9d3b865ed8',
+              currentBalance: '0x77f9fd8d99e7e0',
+              error: '',
+              tokenFees: [
+                {
+                  token: {
+                    address: '0x1234567890abcdef1234567890abcdef12345678',
+                    decimals: 6,
+                    symbol: 'USDC',
+                  },
+                  balanceNeededToken: '0x12C4B0',
+                  currentBalanceToken: '0x4C4B40',
+                  feeRecipient: '0xBAB951a55b61dfAe21Ff7C3501142B397367F026',
+                  rateWei: '0x216FF33813A80',
+                },
+                {
+                  token: {
+                    address: '0x01234567890abcdef1234567890abcdef1234567',
+                    decimals: 3,
+                    symbol: 'DAI',
+                  },
+                  balanceNeededToken: '0xC8A',
+                  currentBalanceToken: '0x2710',
+                  feeRecipient: '0xBAB951a55b61dfAe21Ff7C3501142B397367F026',
+                  rateWei: '0x216FF33813A80',
+                },
+              ],
+            },
+          ],
+          stateDiff: {},
+          feeEstimate: 972988071597550,
+          baseFeePerGas: 40482817574,
+        },
+      ],
+      blockNumber: '0x1293669',
+      id: 'faaab4c5-edf5-4077-ac75-8d26278ca2c5',
+    },
+  },
+};
+
+// Skipped: Detox suite was disabled
+// (https://consensys.slack.com/archives/C02U025CVU4/p1778589879443169).
+// Un-skip after Appium validation on main-e2e.
+appiumTest.describe.skip(
+  SmokeConfirmations('Send native asset Gas Station using EIP-7702'),
+  () => {
+    appiumTest.describe.configure({ timeout: 2500000 });
+
+    const setupCommonMocks = async (mockServer: Mockttp) => {
+      await setupMockRequest(mockServer, {
+        requestMethod: 'GET',
+        url: SIMULATION_ENABLED_NETWORKS_WITH_RELAY.urlEndpoint,
+        response: SIMULATION_ENABLED_NETWORKS_WITH_RELAY.response,
+        responseCode: 200,
+      });
+      await setupMockRequest(mockServer, {
+        requestMethod: 'GET',
+        url: `${LOCALHOST_SENTINEL_URL}/network`,
+        response: SIMULATION_ENABLED_NETWORKS_WITH_RELAY.response,
+        responseCode: 200,
+      });
+      await setupMockPostRequest(
+        mockServer,
+        LOCALHOST_SENTINEL_URL,
+        SIMULATION_GAS_STATION_MOCK.requestBody,
+        SIMULATION_GAS_STATION_MOCK.response,
+        {
+          statusCode: 200,
+          ignoreFields: SIMULATION_GAS_STATION_MOCK.ignoreFields,
+          priority: 1000,
+        },
+      );
+      await setupRemoteFeatureFlagsMock(
+        mockServer,
+        Object.assign({}, ...remoteFeatureEip7702),
+      );
+
+      await setupMockRequest(mockServer, {
+        url: /accounts\.api\.cx\.metamask\.io\/v4\/multiaccount\/balances/,
+        response: {
+          balances: [
+            {
+              object: 'token',
+              address: '0x0000000000000000000000000000000000000000',
+              symbol: 'ETH',
+              name: 'Ether',
+              type: 'native',
+              decimals: 18,
+              chainId: 1337,
+              balance: '10.000000000000000000',
+              accountAddress: `eip155:1337:${DEFAULT_FIXTURE_ACCOUNT}`,
+            },
+          ],
+          unprocessedNetworks: [],
+        },
+        requestMethod: 'GET',
+        responseCode: 200,
+      });
+    };
+
+    const createFixture = ({ localNodes }: { localNodes?: LocalNode[] }) => {
+      const node = localNodes?.[0] as unknown as AnvilManager;
+      const rpcPort =
+        node instanceof AnvilManager
+          ? (node.getPort() ?? AnvilPort())
+          : undefined;
+      return new FixtureBuilder()
+        .withNetworkController({
+          chainId: '0x539',
+          rpcUrl: `http://localhost:${rpcPort ?? AnvilPort()}`,
+          type: 'custom',
+          nickname: 'Local RPC',
+          ticker: 'ETH',
+        })
+        .withDisabledSmartTransactions()
+        .build();
+    };
+
+    const localNodeOptions = [
+      {
+        type: LocalNodeType.anvil,
+        options: {
+          hardfork: 'prague' as Hardfork,
+          loadState:
+            './tests/smoke-appium/confirmations/transactions/7702/withDelegatorContracts.json',
+        },
+      },
+    ];
+
+    appiumTest(
+      'Send ETH via EIP-7702 paying gas with USDC',
+      async ({ driver: _driver, currentDeviceDetails }) => {
+        await withFixtures(
+          {
+            fixture: createFixture,
+            restartDevice: true,
+            localNodeOptions,
+            currentDeviceDetails,
+            testSpecificMock: async (mockServer: Mockttp) => {
+              await setupCommonMocks(mockServer);
+
+              await setupMockPostRequest(
+                mockServer,
+                LOCALHOST_SENTINEL_URL,
+                {
+                  jsonrpc: '2.0',
+                  method: 'eth_sendRelayTransaction',
+                },
+                TRANSACTION_RELAY_SUBMIT_NETWORKS_MOCK.response,
+                {
+                  statusCode: 200,
+                  ignoreFields: ['id', 'params'],
+                  priority: 999,
+                },
+              );
+
+              await setupMockRequest(mockServer, {
+                requestMethod: 'GET',
+                url: `${LOCALHOST_SENTINEL_URL}/smart-transactions/${TRANSACTION_UUID_MOCK}`,
+                response: {
+                  transactions: [
+                    {
+                      hash: TRANSACTION_RELAY_STATUS_NETWORKS_MOCK.response
+                        .transactions[0].hash,
+                      status: RelayStatus.Success,
+                    },
+                  ],
+                },
+                responseCode: 200,
+              });
+            },
+          },
+          async () => {
+            const usdcValues = {
+              fiatAmount: '$3.11',
+              tokenAmount: '1.23 USDC',
+              balance: 'Bal: $12.62 USD',
+            };
+            const daiValues = {
+              fiatAmount: '$8.10',
+              tokenAmount: '3.21 DAI',
+              balance: 'Bal: $25.25 USD',
+            };
+            await loginToAppPlaywright({ scenarioType: 'e2e' });
+            await startRedesignedNativeSendFiveEthToReview(
+              RECIPIENT_ADDRESS_MOCK,
+            );
+
+            await Assertions.expectElementToBeVisible(
+              RowComponents.NetworkFeeGasFeeTokenArrow,
+              { description: 'Gas Fee Token Arrow' },
+            );
+
+            await TransactionConfirmView.tapGasFeeTokenPill();
+
+            await Assertions.expectElementToBeVisible(
+              Matchers.getElementByText('Select a token'),
+              { description: 'Modal is visible' },
+            );
+
+            await GasFeeTokenModal.checkAmountFiat('DAI', daiValues.fiatAmount);
+            await GasFeeTokenModal.checkAmountToken(
+              'DAI',
+              daiValues.tokenAmount,
+            );
+            await GasFeeTokenModal.checkBalance('DAI', daiValues.balance);
+
+            await GasFeeTokenModal.checkAmountFiat(
+              'USDC',
+              usdcValues.fiatAmount,
+            );
+            await GasFeeTokenModal.checkAmountToken(
+              'USDC',
+              usdcValues.tokenAmount,
+            );
+            await GasFeeTokenModal.checkBalance('USDC', usdcValues.balance);
+            await GasFeeTokenModal.tapToken('USDC');
+
+            await Assertions.expectElementToNotBeVisible(
+              Matchers.getElementByText('Select a token'),
+              { description: 'Select a token modal closed', timeout: 10000 },
+            );
+
+            await Assertions.expectElementToHaveText(
+              RowComponents.NetworkFeeGasFeeTokenSymbol,
+              'USDC',
+              {
+                description: 'Gas fee token pill shows USDC',
+                timeout: 15000,
+              },
+            );
+
+            const symbolElementLabel =
+              await RowComponents.getNetworkFeeGasFeeTokenSymbolText();
+
+            await Assertions.checkIfTextMatches(symbolElementLabel, 'USDC');
+            await Assertions.expectTextDisplayed(usdcValues.fiatAmount, {
+              description: 'USDC gas fee fiat amount on confirmation',
+            });
+
+            await FooterActions.tapConfirmButton();
+            await TabBarComponent.tapActivity();
+
+            await Assertions.expectTextDisplayed('Confirmed', {
+              description: 'Activity status Confirmed',
+            });
+          },
+        );
+      },
+    );
+  },
+);
